# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.AllNewProducts:
    properties:
      news_products:
        items:
          $ref: '#/definitions/models.NewProductsList'
        type: array
    type: object
...
  models.CategoryReq:
    properties:
      category_name:
        type: string
    type: object
  models.CategoryUpdateReq:
    properties:
      category_name:
        type: string
      id:
        type: integer
    type: object
  models.FailureInfo:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: string
    type: object
...
  models.MediaLink:
    properties:
      mediaLink:
        type: string
    type: object
  models.NewProductReq:
    properties:
      brand_id:
        type: integer
      camera:
        type: string
      characteristics:
        additionalProperties:
          type: string
        example:
          key: value
        type: object
      description:
        type: string
      dioganal:
        type: number
      display_type:
        type: string
      media links:
        items:
          $ref: '#/definitions/models.MediaLink'
        type: array
      new_price:
        type: number
      old_price:
        type: number
      os_type:
        type: string
      title:
        type: string
    type: object
  models.NewProductRes:
    properties:
      brand_id:
        type: integer
      camera:
        type: string
      characteristics:
        additionalProperties:
          type: string
        example:
          key: value
        type: object
      created_at:
        type: string
      description:
        type: string
      dioganal:
        type: number
      display_type:
        type: string
      id:
        type: integer
      media links:
        items:
          $ref: '#/definitions/models.MediaLink'
        type: array
      new_price:
        type: number
      old_price:
        type: number
      os_type:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.NewProductsList:
    properties:
      brand_id:
        type: integer
      id:
        type: integer
      new_price:
        type: number
      old_price:
        type: number
      title:
        type: string
    type: object
...
  models.SuccessInfo:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
  models.UpdateMediaLink:
    properties:
      id:
        type: integer
      media_link:
        type: string
    type: object
  models.UpdateNewProductReq:
    properties:
      brand_id:
        type: integer
      camera:
        type: string
      characteristics:
        additionalProperties:
          type: string
        example:
          key: value
        type: object
      description:
        type: string
      dioganal:
        type: number
      display_type:
        type: string
      id:
        type: integer
      media_link:
        items:
          $ref: '#/definitions/models.UpdateMediaLink'
        type: array
      new_price:
        type: number
      old_price:
        type: number
      os_type:
        type: string
      title:
        type: string
    type: object
...
  repo.CategoryResponse:
    properties:
      catergoryName:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      products:
        items:
          $ref: '#/definitions/repo.ProductForList'
        type: array
      updatedAt:
        type: string
    type: object
  repo.Empty:
    type: object
  repo.MediaRes:
    properties:
      id:
        type: integer
      mediaLink:
        type: string
      productId:
        type: integer
    type: object
  repo.ProductForList:
    properties:
      brandId:
        type: integer
      categoryId:
        type: integer
      id:
        type: integer
      mediaLinks:
        items:
          $ref: '#/definitions/repo.MediaRes'
        type: array
      price:
        type: number
      title:
        type: string
    type: object
...
host: 100.26.188.92:7070
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is climate_uz server api server
  termsOfService: 2 term climate_uz
  title: Climate_uz api
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: this method renames the category with id
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: Update  category
      tags:
      - Category
  /category:
    patch:
      consumes:
      - application/json
      description: this method renames the category with id
      parameters:
      - description: UpdateCategoryReq
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Update  category
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: this method create new category
      parameters:
      - description: CategoryRequest
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Create new categorys
      tags:
      - Category
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: This function is for deleting categories by Id
      parameters:
      - description: categoryId
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Empty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Delete  category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: this method get category by id
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: Get  category
      tags:
      - Category
...
  /new-product:
    patch:
      consumes:
      - application/json
      description: "this \tmethod update news product"
      parameters:
      - description: UpdateRequest
        in: body
        name: news
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNewProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NewProductRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: "update \tnews product"
      tags:
      - News
    post:
      consumes:
      - application/json
      description: This method add news products
      parameters:
      - description: NewsProductRequest
        in: body
        name: news
        required: true
        schema:
          $ref: '#/definitions/models.NewProductReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.NewProductRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Add news product
      tags:
      - News
  /new-product/{id}:
    delete:
      consumes:
      - application/json
      description: "this \tmethod delete news product by id"
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: delete news product info
      tags:
      - News
    get:
      consumes:
      - application/json
      description: This method get news product info by id
      parameters:
      - description: newproductID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NewProductRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: get news products
      tags:
      - News
  /new-products:
    get:
      consumes:
      - application/json
      description: Through this api can get all new products
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllNewProducts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: get all news products
      tags:
      - News
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"